rag_system_minimal/requirements.txt
# Minimal RAG System Requirements (Assignment 3 Only)

# Core ML/NLP
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# OpenAI API
openai>=1.0.0

# Environment variables
python-dotenv>=1.0.0
